Working on getting correct format to easily turn into vtk - last thing that needs to be addressed is `offsets`. Then need to add functionality to print to JSON correctly.
